Move In, InPipe and InProcessor out of SQL to the common QL project. Move tests classes to the QL project. Create SQL dedicated In class to handle SQL specific data types. Update SQL classes to use the InPipe and InProcessor QL classes. Extract common Foldables methods in QL project.
