+++ x-pack/plugin/sql/src/main/java/org/elasticsearch/xpack/sql/expression/function/aggregate/PercentileRank.java
@@ -13,7 +13,7 @@ import org.elasticsearch.xpack.ql.tree.NodeInfo;
 import org.elasticsearch.xpack.ql.tree.Source;
 import org.elasticsearch.xpack.ql.type.DataType;
 import org.elasticsearch.xpack.ql.type.DataTypes;
-import org.elasticsearch.xpack.sql.expression.Foldables;
+import org.elasticsearch.xpack.sql.expression.SqlFoldables;
 
 import java.util.List;
 
@@ -69,6 +69,6 @@ public class PercentileRank extends AggregateFunction implements EnclosedAgg {
 
     @Override
     public String innerName() {
-        return Double.toString(Foldables.doubleValueOf(value));
+        return Double.toString(SqlFoldables.doubleValueOf(value));
     }
 }
