+++ x-pack/plugin/sql/src/main/java/org/elasticsearch/xpack/sql/planner/QueryTranslator.java
@@ -88,8 +88,8 @@ import java.util.Set;
 
 import static java.util.Collections.singletonList;
 import static org.elasticsearch.xpack.ql.expression.Expressions.id;
-import static org.elasticsearch.xpack.sql.expression.Foldables.doubleValuesOf;
-import static org.elasticsearch.xpack.sql.expression.Foldables.valueOf;
+import static org.elasticsearch.xpack.ql.expression.Foldables.valueOf;
+import static org.elasticsearch.xpack.sql.expression.SqlFoldables.doubleValuesOf;
 
 final class QueryTranslator {
